+++ curvlinops/trace/hutchinson.py
@@ -70,13 +70,11 @@ def hutchinson_trace(
     """
     dim = assert_is_square(A)
     assert_matvecs_subseed_dim(A, num_matvecs)
-    dev, dt = (
-        (A._infer_device(), A._infer_dtype())
-        if isinstance(A, PyTorchLinearOperator)
-        else (A.device, A.dtype)
-    )
     G = column_stack(
-        [random_vector(dim, distribution, dev, dt) for _ in range(num_matvecs)]
+        [
+            random_vector(dim, distribution, A.device, A.dtype)
+            for _ in range(num_matvecs)
+        ]
     )
 
     return einsum("ij,ij", G, A @ G) / num_matvecs